refactor(webapp): split realm helpers
