+++ packages/webapp/src/kernel/realm/helpers/cli.ts
@@ -1,0 +1,83 @@
+import type { ColorApi } from './color.js';
+
+export interface CliDeps {
+  writeStdout: (value: string) => void;
+  writeStderr: (value: string) => void;
+  exit: (code: number) => never;
+  color: ColorApi;
+}
+
+export interface CliDieOpts {
+  exitCode?: number;
+  prefix?: string;
+}
+
+export interface CliWarnOpts {
+  prefix?: string;
+}
+
+export interface CliApi {
+  die(msg: unknown, opts?: number | CliDieOpts): never;
+  out(value: unknown): void;
+  warn(msg: unknown, opts?: CliWarnOpts): void;
+  help(text: string): never;
+}
+
+export function createCli(deps: CliDeps): CliApi {
+  const toLine = (v: unknown): string => {
+    if (typeof v === 'string') return v;
+    if (v instanceof Error) return v.message;
+    if (v === null || v === undefined) return String(v);
+    try {
+      return JSON.stringify(v);
+    } catch {
+      return String(v);
+    }
+  };
+  const formatPrefixed = (
+    color: (s: string) => string,
+    prefix: string | undefined,
+    text: string
+  ): string => {
+    if (prefix === undefined) return `${color('Error:')} ${text}\n`;
+    if (prefix === '') return `${color(text)}\n`;
+    return `${color(`${prefix}:`)} ${text}\n`;
+  };
+  const formatWarn = (prefix: string | undefined, text: string): string => {
+    if (prefix === undefined) return `${deps.color.yellow('Warning:')} ${text}\n`;
+    if (prefix === '') return `${deps.color.yellow(text)}\n`;
+    return `${deps.color.yellow(`${prefix}:`)} ${text}\n`;
+  };
+  return {
+    die(msg: unknown, opts?: number | CliDieOpts): never {
+      const exitCode = typeof opts === 'number' ? opts : (opts?.exitCode ?? 1);
+      const customPrefix =
+        typeof opts === 'object' && opts !== null && 'prefix' in opts ? opts.prefix : undefined;
+      const text = toLine(msg);
+      deps.writeStderr(formatPrefixed(deps.color.red, customPrefix, text));
+      deps.exit(exitCode);
+      throw new Error('unreachable');
+    },
+    out(value: unknown): void {
+      if (typeof value === 'string') {
+        deps.writeStdout(value.endsWith('\n') ? value : `${value}\n`);
+        return;
+      }
+      try {
+        deps.writeStdout(`${JSON.stringify(value, null, 2)}\n`);
+      } catch {
+        deps.writeStdout(`${String(value)}\n`);
+      }
+    },
+    warn(msg: unknown, opts?: CliWarnOpts): void {
+      const customPrefix =
+        typeof opts === 'object' && opts !== null && 'prefix' in opts ? opts.prefix : undefined;
+      deps.writeStderr(formatWarn(customPrefix, toLine(msg)));
+    },
+    help(text: string): never {
+      deps.writeStdout(text.endsWith('\n') ? text : `${text}\n`);
+      deps.exit(0);
+      throw new Error('unreachable');
+    },
+  };
+}
