+++ packages/webapp/src/kernel/realm/helpers/node-tty.ts
@@ -1,0 +1,8 @@
+class ReadStream {}
+class WriteStream {}
+
+export const nodeTty = {
+  isatty: (_fd: number): boolean => false,
+  ReadStream,
+  WriteStream,
+};
