+++ packages/webapp/src/kernel/realm/helpers/time.ts
@@ -1,0 +1,55 @@
+const DURATION_UNITS_MS: Record<string, number> = {
+  ms: 1,
+  s: 1000,
+  m: 60_000,
+  h: 3_600_000,
+  d: 86_400_000,
+  w: 604_800_000,
+  M: 2_629_800_000,
+  y: 31_557_600_000,
+};
+
+export interface TimeApi {
+  parseDuration(spec: string | number): number;
+  ago(spec: string | number, from?: Date): Date;
+  range(spec: string | number, from?: Date): { start: Date; end: Date };
+  future(spec: string | number, from?: Date): { start: Date; end: Date };
+  gmailDate(spec: string | number, from?: Date): string;
+}
+
+function parseDuration(spec: string | number): number {
+  if (typeof spec === 'number' && Number.isFinite(spec)) return Math.trunc(spec);
+  if (typeof spec !== 'string')
+    throw new TypeError('time.parseDuration: spec must be string or number');
+  const trimmed = spec.trim();
+  const m = /^([0-9]+(?:\.[0-9]+)?)\s*(ms|s|m|h|d|w|M|y)?$/.exec(trimmed);
+  if (!m) throw new RangeError(`time.parseDuration: unrecognized spec "${spec}"`);
+  const n = Number(m[1]);
+  const unit = (m[2] ?? 'ms') as keyof typeof DURATION_UNITS_MS;
+  return Math.trunc(n * DURATION_UNITS_MS[unit]);
+}
+
+function pad2(n: number): string {
+  return n < 10 ? `0${n}` : String(n);
+}
+
+export const time: TimeApi = {
+  parseDuration,
+  ago(spec, from = new Date()) {
+    return new Date(from.getTime() - parseDuration(spec));
+  },
+  range(spec, from = new Date()) {
+    const end = new Date(from.getTime());
+    const start = new Date(end.getTime() - parseDuration(spec));
+    return { start, end };
+  },
+  future(spec, from = new Date()) {
+    const start = new Date(from.getTime());
+    const end = new Date(start.getTime() + parseDuration(spec));
+    return { start, end };
+  },
+  gmailDate(spec, from = new Date()) {
+    const d = new Date(from.getTime() - parseDuration(spec));
+    return `${d.getFullYear()}/${pad2(d.getMonth() + 1)}/${pad2(d.getDate())}`;
+  },
+};
